Moving logout functions to common auth lib
