+++ src/views/Settings/Settings.tsx
@@ -21,7 +21,6 @@ import * as data from "data";
 import * as moment from "moment";
 
 import Select from "react-select";
-import ITC from "ITC";
 import { ValidPreference } from "preferences";
 import { Link, useParams } from "react-router-dom";
 import { _, pgettext, interpolate } from "translate";
@@ -35,7 +34,7 @@ import { SpritePack } from "sfx_sprites";
 import { current_language, setCurrentLanguage, languages } from "translate";
 import { toast } from "toast";
 import { profanity_regex } from "profanity_filter";
-import { logout } from "NavBar";
+import { logout, logoutOtherDevices, logoutAndClearLocalData } from "auth";
 import { Flag } from "Flag";
 import { EventEmitter } from "eventemitter3";
 import { LineText } from "misc-ui";
@@ -50,56 +49,6 @@ import { SocialLoginButtons } from "SignIn";
 import swal from "sweetalert2";
 
 export const MAX_DOCK_DELAY = 3.0;
-
-ITC.register("logout", (device_uuid) => {
-    if (device_uuid !== data.get("device.uuid", "")) {
-        swal("This device has been logged out remotely").then(logout).catch(logout);
-    }
-});
-
-function logoutOtherDevices() {
-    swal({
-        text: "Logout of other devices you are logged in to?",
-        showCancelButton: true,
-    })
-        .then(() => {
-            ITC.send("logout", data.get("device.uuid"));
-            swal("Other devices have been logged out").then(ignore).catch(ignore);
-        })
-        .catch(ignore);
-    //get("/api/v0/logout?everywhere=1").then(console.log).catch(errorAlerter);
-}
-
-function logoutAndClearLocalData() {
-    try {
-        get("/api/v0/logout")
-            .then(() => {
-                window.location.href = "/";
-            })
-            .catch(errorLogger);
-    } catch (e) {
-        console.warn(e);
-    }
-
-    try {
-        const cookies = document.cookie.split(";");
-
-        for (let i = 0; i < cookies.length; i++) {
-            const cookie = cookies[i];
-            const eqPos = cookie.indexOf("=");
-            const name = eqPos > -1 ? cookie.substr(0, eqPos) : cookie;
-            document.cookie = name + "=;expires=Thu, 01 Jan 1970 00:00:00 GMT";
-        }
-    } catch (e) {
-        console.warn(e);
-    }
-
-    try {
-        localStorage.clear();
-    } catch (e) {
-        console.warn(e);
-    }
-}
 
 interface SettingsState {
     profile?: any;
